chore: create usage parser file to house usage specific functionality
